backend/requirements.txt
Brotli==1.0.4
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ez-setup==0.9
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
whitenoise==4.1
